+++ spidet/spike_detection/nmf.py
@@ -50,13 +50,13 @@ class Nmf:
 
     def nmf_run(
         self,
-        preprocessed_data: np.ndarray[np.dtype[np.float64]],
+        preprocessed_data: np.ndarray[np.dtype[float]],
         n_runs: int,
     ) -> Tuple[
         Dict,
-        np.ndarray[np.dtype[np.float64]],
-        np.ndarray[np.dtype[np.float64]],
-        np.ndarray[np.dtype[np.float64]],
+        np.ndarray[np.dtype[float]],
+        np.ndarray[np.dtype[float]],
+        np.ndarray[np.dtype[float]],
     ]:
         """
         This method performs a number of nonegative matrix factorization runs for the rank
@@ -66,7 +66,7 @@ class Nmf:
 
         Parameters
         ----------
-        preprocessed_data: numpy.ndarray[numpy.dtype[numpy.float64]]
+        preprocessed_data: numpy.ndarray[numpy.dtype[float]]
             The data used as input for NMF algorithm.
 
         n_runs: int
@@ -75,9 +75,9 @@ class Nmf:
         Returns
         -------
         Tuple[Dict,
-        np.ndarray[np.dtype[np.float64]],
-        np.ndarray[np.dtype[np.float64]],
-        np.ndarray[np.dtype[np.float64]]]
+        np.ndarray[np.dtype[float]],
+        np.ndarray[np.dtype[float]],
+        np.ndarray[np.dtype[float]]]
             A dictionary containing the rank, the minimum reconstruction error, the cophenetic correlation
             and the instability index, together with the consensus, :math:`W` and :math:`H` matrices.
         """
